Increased Qaz's cloud generation speed
Clouds spawn 3-5 at a time and bloom on a 5x5 tile (24 tiles), excluding the player. When piety is 160 or higher, the number of sustained clouds averages 7-8. In theory, the more enemies there are, the more likely the clouds are to hit an enemy. In practice, however, most players prefer smaller 1v1 or 1v3 engagements. The fewer enemies a player faces, the less likely the clouds are to hit them accurately. The return seems too unreliable for the risk of not spawning clouds in loud noises and corridors. Therefore, I would like to increase the spawn rate of clouds to increase the number of clouds maintained to an average of 9 (Piety 160).