Relax criteria to allow Qaz's clouds to be generated in complex terrain
Qaz's clouds don't spawn in corridor-type terrain, which is probably a device intended for fighting in open areas. However, the current criteria is too strict and prevents clouds from spawning in ambiguous places that are hardly corridors as the terrain gets more complex. Therefore, I'd like to relax the criteria to still prevent clouds from spawning in corridors (2 tiles or less of space for the player to move), but allow them to spawn in other terrain.