Reduce the cost of Upheaval
Upheaval is a decent skill, but you need to use it 2-3 times to reliably take out an enemy, and in situations where you need to use it more than 3 times, Disaster Area can reliably do a similar job in 1 turn. Therefore, I would like to reduce the cost slightly to make the upheaval a bit more competitive.