Gell's Gravitas -> Gell's Gavitambourine
Gell's Gravitas has been in an awkward place for many years. It's
conceptually cute, but generally agreed to be one of the weakest spells in
the game, and there are implementation issues with making how it works
transparent to players (eg: beam paths often mean that monsters the player
expects to get pulled do not, and even if they do, they may move
immediately afterward, making it *appear* as though they did not). I tried
rewriting it in various ways, but it always felt like it was stepping on
the toes of another spell in the same school that did that things better.
IMB and Dispersal are both much better at getting things away from the
player, Piledriver is (hopefully!) a more direct and intuitive way to slam
monsters into walls for value, and Gavotte can do that at a larger scale.

However, by turning it into an evocable effect, we may be able to lean into
the 'gather monsters into one place' facet more effectively than one could
from a repeatable effect, as well as improving direct damage without
concerns that it's in the wrong school for that. Hence, the tambourine!
(Because I also cannot resist a silly portmanteau)

The tambourine is smite-targeted in a large radius (that scales with evo)
and tries to pull every monster in that radius into as small an area as
possible, then holds non-friendly ones in place there for several turns, as
well as dealing some immediate (non-collision-based) damage to them.

It currently alternates with phial of floods as a 'control' evoker, though
once the pool of evokers expands just a little bit more (which I'm still
hoping to do in 0.32), I'd like to remove these '1 of 2' pools from evokers
and just pull 5 at random from the entire list. I think it's better for run
variety and I don't see the need to guarantee an even distribution of
evoker roles (or even power) from run to run in the same way that it makes
sense for wands (since those are a bread-and-butter early toolset).

(I also realize we've never had proper names in evokers before, but at this
point they're *basically* unrands - coloring and all - so it's probably
fine, right?)

The tile is... very placeholdery, but I tried my best. >.>