Fix two subtle issues with stats panel weapon colour in WebTiles
Two related issues existed, relating to the text colour of the displayed
weapons in the stats panel (top-right corner) in WebTiles:
  1. Due to these colours being set on the JS side, which only has
     access to the inventory set of menu_colour options, the setting
     menu_colour += stats:<colour>:<weapon description> was essentially
     useless, as the inventory setting was always used.
  2. The item colour for the stats-weapon panel was not updated when a
     weapon was equipped, meaning it only displayed in the correct
     (equipped, typically lightgreen) colour when reloading the game, or
     when equipping the weapon directly from the ground.
Since the expansion of quiver functionality, these issues have not
existed for the quiver slot, as the entire quiver description including
the colour is computed on the C++ side. Hence (particularly with
launchers) you could end up with the same item in your weapon and quiver
slots written in different colours.

Fix both issues by passing the weapon colour(s) directly from the C++
side as a separate piece of information. Incorporate the old
unarmed_attack_colour information into this.