Another WebTiles weapon display fix (Ge0ff)
The code previously assumed that knowing the base type, sub type,
enchantment, quantity, flags, and (player) inscription of an item
was enough to fully identify it, so an update was only sent to the
webserver if one of these things changed.

This meant that if you happened to have two randart weapons of the
same base type and enchantment, by adjusting weapon slots with =i..
it was possible to trick the display into showing the complete wrong
weapon description to the weapon you actually had equipped, and this
would persist forever unless one of the weapons somehow got modified.

Fix this, and hopefully any future issues with anything like this,
by checking the name of the item. It really ought to be the case that
the full name of an item (including its randart inscription) should
uniquely identify it, or if not, the displayed descriptions will be
indistinguishable anyway so it doesn't matter.