Replace a couple uses of Blink Away with Blink Ranged
Blink Away will happily move the monster out of sight of their target
(which is usually the player) while Blink Ranged attempts to create
distance while still being able to see the target.

It seems better behaviour in the majority of cases if enemies don't leave
the range at which they can threaten the player altogether. I left Blink
Away alone for summoners (who often benefit actively from leaving their
summons to hit you, even if they can't see you themselves) and boulder
beetles (because they can be so dangerous early on and them sometimes
disappearing is a more of a mercy to the player than a nuissance)