fix: disable requirejs timeout altogether (mumra, ge0ff)
The default value for this timeout is 7s. Since tilesheets are sequenced
before js modules in client loading (because they are loading from
html), on a slow connection this timeout can be hit for non-cached
tilesheets. I didn't find a reasonable version-independent way to defer
tilesheet loading without a lot of refactoring, so instead, just disable
the requirejs timeout. It would be possible to set this to a large
value, but from the user's perspective, since we already have a loading
screen, there's no real difference. It's just a matter of whether the
errors show up in the developer console or not, but the hang-on-load
behavior is the same either way. I also couldn't see any value in making
a user-facing timeout error on the load screen. So, this commit takes by
far the simplest approach and just disables the timeout.

For reference, in 3G throttling I was seeing 40-70s load times, and in
4G throttling this can go up to 30. I'm not sure we are trying to
accommodate 3G, but 4G might be reflective of bad connections we would
want to try to accommodate.