fix: further tweaks to webtiles lazy image loading
In testing, the first attempt didn't look very good on a live server,
because the image load was very visible. This commit changes the
approach to randomly select and ensure loading for an image every time a
logged-in player enters the lobby. Like the prior attempt, it forces
loading by removing the lazy attribute; as the comment notes, I'm not
sure the standards require a browser to react to this, but they do seem
to in practice.