Magnavolt (level 7 Air/Earth spell)
An attempt to fill out more cross-over spells, since I think those can
present interesting skilling decisions, as well as continuing to add more
options for high-level casters in general. Magnavolt is a smite-targeted
spell that coats a target with a thin layer of magnetised metal and then
fires a bolt of electricity at every visible enemy that is magnetised in
this way. (Meaning that the first time you cast it, it will fire one bolt,
but repeatedly casting it will fire 2, 3, 4, and so forth).

The magnetised status is short-lived, reduces the target's EV to 0, and
also causes them to leave behind an electroferric vortex for a brief while
after death (which is stationary, but perma-magnetised, so that it can
continue to serve as an anchor for the spell if the initial target
immediately dies, which helps to keep the ramp-up more predictable in
situations where the caster is actually successfully killing things).

By blending the two opposing schools together, this spell offers Air more
controllable (and less unsafe) long-range damage at this level range, while
giving earth better AoE (in exchange for it being more resistable) with a
simple unique twist. Ideally this tempts some people to pivot in ways that
previously weren't as rewarding, while offering a new simple damage option
for high level mages.