Fulsome Fusillade: level 8 Conjurations/Alchemy
An attempt at giving alchemy a high-level blasty spell that feels
sufficiently distinct from those which already exist, while also feeling
appropriate to the thematic spirit of the school.

Fulsome Fusillade gives the player a status which causes beakers of random
reagents to rain on nearby targets for each of the next 5 turns, causing
3x3 explosion of either fire, cold, electricity, or poison arrow aimed on
top of 3 random enemies. But importantly, if any of those explosions
*overlap*, the area of overlap is instead hit by a volatile alchemical
reaction that deals a larger amount of non-elemental damage and has a
chance to apply a random debuff.

In effect, one can consider it somewhere between Polar Vortex and Ignition;
it does mp-efficient damage over time but benefits greatly from targets
being adjacent to each other. It can work in corridors and hit enemies from
a far range, but becomes far less useful against either single targets OR
a very large number of targets - being ideal against a single clustered
group that will focus the random hits into a small area.

Many aspects of this spell are still tentative and subject to further
adjustment as playtesting happens.