Rimeblight: New Level 7 Ice/Necromancy spell
A beam-targeted irresistable enchantment that inflicts Rimeblight upon a
living/holy/demonic creature. The creature will take AC-ignoring cold
damage over several turns as ice devours them from the inside out, and also
periodically explode in a flurry of ice shards that damages nearby hostiles
(but not the player or their allies!) and can spread the disease to
succeptible creatures it hits.

Creatures with rimeblight on them that drop below 20% hp (ie: the 'almost
dead' threshold) will instantly die and be replaced with a short-lived
pillar of rime. (Purely decorative, but the flavor is too good to pass up
here, in my opinion)

Beyond the neat flavour and unusual spreading mechanics, this provides a
high level offensive necromancy option that isn't ally-based (and also a
little more variety in high-level ice magic in general).