Seismic Cannonade (level 7 Alchemy/Earth)
Another attempt to fill in some high-level alchemy options of a very
different style than Fulsome Fusillade, this spell consturcts 3 seismic
cannons near the caster which are stationary turrets that fire medium-range
projectiles each turn and have a built in 'super' they can charge up for.

The cannons are initially created with 1/2 of their max hp, but heal a
small amount with each shot they fire (as they dig up and incorporate more
material from the earth into their frames). Once they reach full hp, the
caster can recast the spell to unleash a seismic shockwave that does
significant area damage to a targeted location and also destroys the
remaining cannons in the process.

This is intended to lean into alchemy's thematics of 'creation' (as
Fusillade leaned into 'chemical reactions'), while providing a high-level
ally spell that feels distinct from, say, a Spellforged Servitor using
bombard or crystal spear (which Earth could already access at this level
with a splash of a secondary school). By contrast, cannonade is noisier,
less flexible with positioning, and requires some defending, but can switch
between wide-scale AoE and single-target damage on the fly. (It was also
an attempt to make another 'modal' ally spell after the positive response
to revamped Call Canine Familiar).

As usual, all of the various numbers involved in this are still provisional
and may need adjusting. (Also: the cannon tile is very sad; I apologise.)