Some tweaks to auto-aiming to make Magnavolt work sanely with autofight
This repurposes the still-implemented but currently unused prefer_farthest
member of direction_chooser and adds spflag::prefer_farthest to let spells
use this (currently just Magnavolt).

It also fixes some bugs I discovered with its implementation in the
process, like how it choose the 'farthest' target by starting with the
default closest target and then cycling backwards once so that it would
'wrap around' to the farthest. Which didn't actually work properly in many
cases where the default target *wasn't* the closest to start with.

Additionally, this adds a preferred_aim member to targeter (which normally
defaults to returning valid_aim) for situations where it is clear that a
certain valid target would not be what the player would *actually* want to
aim at in the vast majority of cases. Default target selection is now done
in two passes - looking for a preferred_aim first, and then picking a
valid_aim second if no preferred_aim is found. This isn't intended to use
any complex heuristics to guide the player, but for simple situations where
'closest valid target' doesn't actually represent good default behaviour.

There may be other spells for which it makes sense to use this, but for now
it's also just used by Magnavolt (otherwise, ranged autofight uses the
spell so badly that you essentially could only ever manually cast it)

For the record: the deep guts of direction_chooser are scary and deeply
unpleasant to work with. I am not 100% confident that what I have done here
won't have some undesired side-effect, though in basic testing other
things seem to still work properly. Fingers crossed. (I also wasn't
thilled to add a member to targeter for just one spell here, but wasn't
sure of a cleaner way to accomplish the goal with what I was working with.)