Don't let allies reduce damage enemies take from Ozocubu's Refrigeration
It always felt both a flavor mismatch *and* unfortunate mechanics that
simulaca (which are immune to Refrigeration) still had anti-synergy with
it by reducing the damage enemies would take from it. Rather than
specifically exclude icy monsters from this calculation (and invite
confusion/debate about what is sufficiently un-warm), only give the huddle
bonus from monsters *allied* with the target taking damage. So massed
enemies take the same reduced damage as now, but your summons will no
longer deign to give them a friendly cuddle for warmth.

This effectively buffs both spells a little, but Refridgeration doesn't
seem so strong these days that it can't handle that, and it feels like a
cute and intuitive synergy.