Make -move and stasis block the self-move portion of Gavotte
You already can't move stationary enemies this way, so it feels like it
should work the same way for players. But unlike many transloctions that
move the player, you're clearly affecting the *general area* and not
specifically yourself, so it seems intuitive that you can still hurl
enemies around even if you remain in the same place.

This does mean that formicids can do things with it that other species
cannot, but it's not clear to me that the advantage of 'can toss enemies
out of your LoS sometimes' outweighs the downside of 'can't ever use it to
rapidly reposition yourself', and formicids aren't exactly a concerningly
powerful species at present. We'll see if that holds up in practice.

(Deliberately miscasting translocation spells to get this property on other
species seems almost certainly worse to attempt than doing something else
with your time, and if someone wants to use !lig to take advantage of this,
then they can be my guest)