Don't double-miscast when miscasting Seismic Shockwave (NormalPerson7)
This was a consequence of how the second spell was embedded inside the
cast code of the first. Instead, directly switch which spell is being
cast inside your_spells, depending on the cannon's mode.

This feels a little ugly, but is at least clear in what it's doing and
seems to solve the original problem without creating new ones.