Make flashes consistent in webtiles
Webtiles only had a full-screen flash but in desktop it was per-tile (with
an optional hitfunc). The colour was already available on screen_cell_t so
it was easy to just send this value in the cell json only if it changed,
and this even saves redraws as we don't need to invalidate the entire map
on a flash, only the changed tiles. And it's much easier on the eyes to
boot.