Refactor / unify how monster passive enchantment auras work
Currently this combines torpor snail slow aura and martyred shade injury
bond aura into a single new system (which will make it very simple to add
similar effects with less duplicated code).

Functional changes:
 -Torpor snails now (deliberately) work through grates/statues. At least
  one vault already assumed that they did, and it doesn't feel like there's
  a good thematic or mechanical reason for them not to
 -Martyed shade's injury bond now wears off if they spend more than a
  couple turns out of LoS of what they're protecting (previously, LoS only
  mattered to apply, and it would otherwise never wear off).
 -Shades now apply injury bond each turn without needing a visible foe
  nearby. (The old behavior could sometimes lead to confusing situations
  where they wouldn't bond an ally for no obvious reason)

It is *possible* there are some minor changes with timings of when effects
are applied or removed, but I think those should be the main effects.