Add marrowcudas to pre-Lair D
These were intended as low-level Yred fodder (who is still using
necrophages for early Black Torch summons, despite those not even spawning
anywhere anymore) as well as an earlygame introduction to a mechanic used
on a much latergame monster, though *that* monster isn't implemented yet
and it's possible its design will end up changing before it is.

Nevertheless, this is a tiny bit of additional variety, another monster
that could be used as a boss in a Sewer or Ossuary, and provides an
earlygame summoner (of which there are very few which aren't uniques) and
very strongly hilights the 'kill the summoner first' approach.

They are a fast, relatively low-damage, flying skeletal enemy with
AF_SWARM, which summons an additional copy of *itself* whenever it deals
damage to a target. (And then those copies can summon other copies, and
so forth, up to a cap of 8). It can block off retreat in corridors and
quickly create overwhelming numbers if ignored, but dies fairly easily when
focused directly.

They spawn (relatively rarely) at a slightly shallower depth than Wights,
as well as sometimes in the early water table (instead of 100% electric
eels all the time.)

I'm still uncertain if their placement is too close to wights and would be
better at some other depth, with other stats. I'll reassess this after some
playtesting.