Add Undying Armouries, a Vaults out-of-depth monster (regret-index)
These replace tentacled monstrosities from the Vaults spawn list entirely,
reducing the number of different zones they appear in, while offering a new
sort of threat that is a strong thematic fit with Vaults itself.

Undying Armouries are a vortex of weapons, animated by a poltergeist formed
from the souls of numerous smiths and weaponmasters. They have decent melee
and two unique spells:

Flashing Balestra is a mini iron shot, usable only from range, that spawns
a short-lived dancing weapon near its target - as one of the souls in the
maelstrom leaps out to duel with you.

Bestow Arms temporarily grants copies of a random summoned weapon to 5-7
nearby allies. These tend to be top-tier items, with good brands and
plusses, with a slight bias towards polearms, but also including launchers!
This is a unique vector of enemy buff that still ties into existing game
systems (and giving an entire band, say, speed longbows or vampiric glaives
ought to be felt!)

The implementation of this was tricky, and while I think I have caught and
fixed most situations where this could devour their real weapon
permanently, I'll be keeping my eye on it for other weirdness.