Exile Arachne from Spider, replace with a new formicid unique
Since the introduction of jorogumos to Spider, Arachne's main gimmick has
felt a little crowded out. They're perfectly good monsters, but she's more
or less just a larger one that blinks, and it feels like she'd be more
noteworthy if she was simply somewhere else, so let's move her.

Arachne now comes with 2-3 orb spiders and can spawn in Vaults:1-4, Elf,
and Depths:1 *only in games where Spider was not generated*. You either get
joros or her, but not both. (Orb spider backup in open terrain also tends
to make her ensnare gimmick more sharply threatening, which is sort of
neat)

Replacing her in Spider is a long-awaited formicid unique. Xak'krixis is
an alchemist on expedition from an isolated formicid kingdom, studying
the fauna of Spider. They cast venom bolt, ignite poison, and an improved
version of fulminant prism which places two at once *and* has no friendly
fire. A straight adaptation of the player spell wouldn't play very well,
but multiple 'safe' bombs can make it less trivial for the player to
simply ignore, while also not endangering the many bands in Spider. (The
noise of them is even a danger of its own down there).

Xak comes with a venom-branded two-hander and a shield (why have a formicid
if they aren't going to use a shield with a two-hander, right?) and can
also dig while moving - but *not* while wandering (to avoid some of the
old boring beetle problem of loudly digging out a level whenever they'd
gotten woken up). Mostly this means they'll only do it while actively
pursuing the player, but also while running away (and possibly in some
other situations).