Fold Lorocyprocas into Hell Beasts and rename/tweak (regret-index)
As per the previous commit, this removes loros and gives their antimagic
attack to hell beasts, which are additionally renamed to 'sin beast' (it
was always a little weird that something which commonly spawns in Pan but
NOT in Hell was named 'hell beast')

They have the same base damage as before, with antimagic on their bite
attack. Their HD was raised from 7->12 (as this directly affects their
antimagic strength) which is still a bit lower than loros, but they're
much faster.

Out of some concern that Geryon might be a little too oppressive for
characters that aren't meaning to tackle the rest of Hell (since horn
troves are fun and should be encouraged!), I reduced the summon cap on
sin beasts from 8->5, and reduced their duration and his cast rate a little
bit. A pure buff to evoking the Horn of Geryon itself also seems unlikely
to be too good?  We'll see how much any of that feels like a problem.