Make Aura of Brilliance no longer an aura
In theory, this spell was supposed to continuously grant brilliance to
allies near the caster, aura-style. In practice, it was kind of broken.
The code looks like it was *attempting* to renew the brilliance duration
each turn, but did not actually do so (reapplying full duration only after
it completely wore off of someone). Also, each application of brilliance
lasted approximately equally long (give or take randomization) as the
entire aura itself. So it was very common for the 'aura' to run out before
the first application of a buff did, and never extend its duration at all.
And then sometimes, it would last *just* long enough for the buff to wear
out and reapply it, effectively doubling the duration.

To make matters more confusing, the aura would immediately turn off if
ever a single turn passed without a buffable ally being in LoS of the
caster. (And there was no icon indicating an aura was even active...)

It would be possible to make it function more like an aura, but I'm not
sure there's a great mechanical benefit to doing it in this case. So
instead, just make it a one-time AoE buff (with about 75% longer duration
than it has at present) and remove the other supporting code.