Fix a rare Gozag crash
When an enemy is killed in a way that causes their corpse to explode (eg:
mindburst or iood) while worshipping Gozag, multiple piles of gold are
scattered around the general area. The amount of gold in each pile is
determined by randomly rounding 'total gold' / 'number of piles'. For
medium-sized creatures specifically, if the player rolled the lowest
possible amount of gold and the highest possible amount of chunks, it was
possible for this to round down to 0 gold in a pile.

Which was already specifically guarded against! Piles of 0 gold will not be
placed by _explode_corpse. ...but _explode_corpse operates using a
reference to the 'original' corpse and while place_monster_corpse attempts
to destroy the corpse item after an explosion, destroy_item() specifically
doesn't do anything to an item with a quantity of 0, so the original corpse
(now gold) would remain with a quantity of 0.

Mostly this was invisible, but upon leaving the current floor, attempting
to save the present floor would trigger an assert for the 0-gold pile being
invalid, causing this crash.

Now just make sure that none of the piles ever has less than 0 gold in it.