Let Flash Freeze damage targets that already have Frozen status
It's a somewhat odd and very specific exception (that does not apply to
any other effect that applies Frozen, either) that Flash Freeze cannot deal
any damage to a player/monster that already has frozen.

Rime Drakes' breath cooldown already made them unlikely to be able to
attack the player any more often than this and marginal buffs to other
users seem low impact.

Admittedly, this makes it possible that a player facing *multiple* rime
drakes at once can get breathed on a lot more often, but it was honestly
kind of weird that they couldn't.