Monster version of Dazzling Flash
Now we can blind the player in an interesting way, dazzling flash can be
unlocked for player ghosts.

The resist mechanism (chance depending on player XL vs spell power) is
perhaps not great and should be reviewed.

I gave the spell to Moon Trolls slightly arbitrarily as I was testing with
them and it felt thematic so I left it in (it's a very rare encounter
anyway).