Increase the effectiveness of the Frozen status
Frozen was applied as a side-effect of Flash Freeze and Creeping Frost,
but didn't have very much impact. It lasted 2-4 turns and increased move
delay by 3 aut, meaning that even if you moved every one of those turns you
were penalized, you'd lose less than 1 turn on average. In most situations,
I think it barely affected how the player should conduct themselves. If a
debuff is going to exist, I think it should be easier to notice that it did
something, so let's try ramping it up slightly.

Frozen now lasts 3-5 turns instead, and increases move delay by 50%
(identical to being slowed). This is fairly conservative, but it's a nice
round number with existing precedent. Frosthearts are already fairly
scary as-is, so I've nudged their HD down by 1. This may not turn out to be
necessary.

As for monsters, Frozen was even more sad. Mostly this didn't matter, as
circumstances where it was applied to monsters was few (and Upheaval is
likely to simply kill them rather than have you worry about how slow they
move now). But the combination of 'Almost always lasts 2 turns' and
'Increases move delay by 4 aut' meant that it was nearly mathematically
impossible for a normal-speed enemy to visibly lose even a single turn
from it. (It being a flat +4 to move cost also had the strange effect
where enemies with fast movement but normal overall speed suffered a
greater penalty than enemies with equivalently high speed for all actions.)

Since debuffs like this tend to need a *larger* impact to notice on
monsters than on players, let's try making it double their move delay
instead (and last slightly longer). This will be relevant for the very
next commit.