Hellfire Mortar (level 7 Fire/Earth spell)
A replacement for seismic cannonade from the other direction. This spell
creates a line of temporary lava in a chosen direction (even through walls
in that direction) and spawns a mortar that moves along that line of lava,
firing autonomously at nearby targets while it does so.

The mortar fires bolts of magma at up to 2 different random enemy targets
each turn, then moves 1 tile further along the path that was carved out for
it. When it reaches the end of this path (or a monster blocks its attempt
to move), it will disappear.

Given that it acts entirely autonomously over a short duration, I have
decided that it is closer to a foxfire or battlesphere than it is to a
lightning spire in terms of whether Okawaru forbids it or Vehumet supports
it. I could see arguments in the other direction, though.

Also: enough years have passed since Hellfire was renamed Damnation that it
is safe to use for something else again, right? (And this time, it really
*is* doing fire damage...)