Improve behavior if player ends up in a wall due to terrain changes
Instead of instantly randomly teleporting the player, try to simply shove
them out of the way of the wall first (and, failing that, try to move them
somewhere in los range first).

There were probably other situations this could previously happen, but it's
easier to do deliberately now because of Hellfire Mortar. Now the player
should almost always be gracefully pushed backwards instead of being able
to trigger an instant teleport this way.