Tweak Rimeblight a little more heavily
Increased the infection chance from rimeblight explosion from 1/2 to 2/3.
Monsters that die to rimeblight now also have the same chance of spreading
the disease to adjacent creatures (without accompanying damage) to avoid
the awkward situation of you otherwise wanting to keep weak things alive
long enough for them to become contagious (without adding death explosions
that risk making it feel too adjacent to inner flame).

You can no longer aim the spell through some of your allies (this was
always an oversight).

More significantly: explosions and disease spread no longer happen out of
line of sight. Previously, the most powerful use of this spell was to
unleash it on a group and then get the heck away so that they would slow
die to each other while the player was safely away. While we do allow some
high level spells to have this capability, it feels like there's a lot more
risk baked into, eg: Discord, to counterbalance this. I'm a bit
uncomfortable with this being the primary use-case for Rimeblight, and want
it to feel sufficiently attractive to use while standing around for a
fight. (The internal ice damage still occurs while out of sight, in the
same way that poison still works out of sight)

I also increase explosion damage somewhat (though note that the increased
rate of spread is fairly significant).

We'll see how this updated version feels to use.