Fix Repel Missiles expiring off of monsters (ge0ff)
The repel missiles status was recently made to decay with time (as part
of the aura rework and polterguardians). I had assumed as the time that
the permanent repel missile status used by Deep Elf Archers and co was
already set to INFINITE_DURATION, but it was not, and now unintentionally
expired with time.