Adjust fugue to level 2
While FUGUE's max stack effect (negative energy attack) is fun and useful, it's not easy to build up to 7 stacks in a real-world combat setting, as the average player won't put themselves in a situation where they're fighting 7 creatures in a row unless they have to. Also, it's not uncommon to have no or very little left at the end of 7 stacks. Therefore, I'm lowering the max stack of this spell to 4 to make it easier to see the final effect. I'm also lowering the level to 2 to reflect the weakened effect.