Reaver background rebalancing
There are two reasons why I want to turn the reaver into a de facto necro-warrior.
1. reaver is a warrior-mage class, but there is no synergy between weapons and magic, they simply mechanically use weapons and magic together.
 * Cinder Acolyte: I start with a flame-branded weapon. Scorch spells can be used to create synergy.
 * HexSlinger: Training with weapons makes it easier to hit enemies with Inner Flame debuffs for the intended effect; training with magic, on the other hand, makes it easier to inflict debuffs on enemies. Either way, there are synergies.
 * Enchanter: Practicing weapons increases your stab damage. Practicing magic makes it easier to assassinate enemies. There is a synergy between practicing either.
However, the reaver has no such synergy, so I believe the reaver's spell list should be changed to allow weapons and magic to synergize with each other.

2. i think that the backgrounds of necromancers and necro-warriors should be separated from each other. necromancers are mages, not warrior-mages. however, their spell composition is a mix of both types. therefore, i would like to see the necromancer's spell list changed to be more mage-like and less warrior-mage-like.

The Reaver needs to change, and the Warrior-mage spells that the Necromancer has need to find a new job, so I decide that it would be a good idea to rebalance the Reaver to Necro-Warrior.