Necromancer background rebalancing
As a follow-up to the Reaver background rebalancing, we're changing the Necromancer's spell list to make it more appropriate for Mages. We wanted to give them another offensive spell that doesn't require melee instead of Vampiric Draining, but unfortunately we couldn't find anything like that in the entry-level necromancy. So for now, I'll leave Vampiric Draining alone.