Properly ascribe blame for deaths due to exploding monsters
This was broken since 3dd439832fe0f5fe295caad1a2fe3e479a757cda which caused
all ikiller information for deaths caused by exploders (ie: ball lightning
or fulminant prisms) to be lost. This isn't conceptually trivial to fix,
as now the monster responsible for the explosion will already be cleaned
up by the time the explosion happens (and the entire blame chain was stored
inside that monster).

The fix I've settled on is to maintain a very transient cache for copies of
monsters that are about to explode. If a scorefile_entry tries to look up
a monster by mid and fails, it will look at this cache instead so that it
can find the dead monster's info as it was when the final_effect was
first scheduled.

(This cache is cleared as soon as fineffs are done processing, so there
shouldn't be any need to worry about its state.)

It's possible there is some other class of fineff that could benefit from
using it, for similar reasons, but this wasn't immediately obvious to me.
However, this should once again let us see whose fault these deaths are.

(Fun fact: I'd thought that Xak'krixis was wildly underperforming, but
apparently a full 90% of their kills weren't being attributed to them. Poor
Lom also lost out on ~30% or so of their kills since the original commit).