Wiglaf has returned for his hat (and brought a very big gun)
Along with the return of playable mountain dwarves, it seemed like a good
joke to bring back an iconic mountain dwarf unique at the same time, just
so long as there was something interesting for Wiglaf to *do* and I think
I may have found that.

Wiglaf can now cast Hellfire Mortar, a thematically appropriate (and
hopefully fun) spell. They've lapsed a little in their worship of Okawaru,
losing Haste to keep them from overlapping with modern Donald too much (but
still know Might). They always spawn with a good randart hat and since,
well, everyone knows Joseph *took* that hat, they alternate with him as a
unique pair.

A tiny bit of their dialogue was restored from the old days, but after
trimming out the cliche drunkenness and vague racism, the vast majority of
it is new. I did my best to manage scottish eye dialect in a similar style;
not sure if I have gone overboard or not.

Implementation note: it's impossible in a conventional way for a monster to
have a hat and a shield at the same time, even using the 'usual' hack used
for Nikola's gloves. So I have shoved Wiglaf's hat into his jewelry slot,
which I dislike rather a lot. But I'm reasonably sure that it works
properly.